+++ frontend/SchedulingAppointmentsWindow.py
@@ -7,14 +7,13 @@ from frontend.ui.assets.files.GLOBALS import *
 from frontend.ui.assets.files.NAVIGATION_FUNCS import *
 from frontend.ui.assets.files.SCHEDULING_LISTENERS import *
 
-import urllib
-import sqlalchemy
+from abstract_main_window import AppointmentMainWindow
 import backend.private.data_manager
 import sys
 
 
 
-class UI(QMainWindow):
+class UI(AppointmentMainWindow):
     def __init__(self):
 
 
@@ -24,128 +23,13 @@ class UI(QMainWindow):
 
         # Session for connecting to the Database
         self.session = backend.private.data_manager.DataManger().session
-
-        # Functions
-        # Frame functions
-        def hideAllFrames():
-
-                # All frames have a height of 681
-                self.InputsFrame.setFixedHeight(0)
-                self.CancelAppointment_Frame.setFixedHeight(0)
-                self.RescheduleAppointment_Frame.setFixedHeight(0)
-
-                # Re-enabling the frame btn togglers
-                self.DisplaySchedule_Btn.setEnabled(True)
-                self.DisplayCancel_Btn.setEnabled(True)
-                self.DisplayReschedule_Btn.setEnabled(True)
-
-                # Restyling frame btn togglers
-                self.DisplaySchedule_Btn.setStyleSheet("QPushButton {\n"
-"    border-image: none;\n"
-"    background-color: #6ECCAF;\n"
-"    color: black;\n"
-"    border: 2px solid black;\n"
-"}\n"
-"\n"
-"QPushButton::hover {\n"
-"    background-color: rgb(139, 231, 100);\n"
-"    color: white;\n"
-"}")
-                self.DisplayCancel_Btn.setStyleSheet("QPushButton {\n"
-"    border-image: none;\n"
-"    background-color: #6ECCAF;\n"
-"    color: black;\n"
-"    border: 2px solid black;\n"
-"}\n"
-"\n"
-"QPushButton::hover {\n"
-"    background-color: rgb(139, 231, 100);\n"
-"    color: white;\n"
-"}")
-                self.DisplayReschedule_Btn.setStyleSheet("QPushButton {\n"
-"    border-image: none;\n"
-"    background-color: #6ECCAF;\n"
-"    color: black;\n"
-"    border: 2px solid black;\n"
-"}\n"
-"\n"
-"QPushButton::hover {\n"
-"    background-color: rgb(139, 231, 100);\n"
-"    color: white;\n"
-"}")
-
-                # Clearing all the fields
-                clearInputs_SA(self)
-                clearInputs_RA(self)
-                clearInputs_CA(self)
-
-        def displayInputsFrame():
-
-                hideAllFrames()
-                self.InputsFrame.setFixedHeight(681)
-
-                # Disabling the toggler btn
-                self.DisplaySchedule_Btn.setEnabled(False)
-                self.DisplaySchedule_Btn.setStyleSheet("QPushButton {\n"
-"    border-image: none;\n"
-"    background-color: rgba(110, 204, 175, .2);\n"
-"    color: rgba(0, 0, 0, .2);\n"
-"    border: 2px solid rgba(0, 0, 0, .2);\n"
-"}\n"
-"\n"
-"QPushButton::hover {\n"
-"    background-color: rgb(139, 231, 100);\n"
-"    color: white;\n"
-"}")
-        def displayCancelAppointmentFrame():
-
-                hideAllFrames()
-                self.CancelAppointment_Frame.setFixedHeight(681)
-
-                # Disabling the toggler btn
-                self.DisplayCancel_Btn.setEnabled(False)
-                self.DisplayCancel_Btn.setStyleSheet("QPushButton {\n"
-"    border-image: none;\n"
-"    background-color: rgba(110, 204, 175, .2);\n"
-"    color: rgba(0, 0, 0, .2);\n"
-"    border: 2px solid rgba(0, 0, 0, .2);\n"
-"}\n"
-"\n"
-"QPushButton::hover {\n"
-"    background-color: rgb(139, 231, 100);\n"
-"    color: white;\n"
-"}")
-        def displayRescheduleAppointmentFrame():
-
-                hideAllFrames()
-                self.RescheduleAppointment_Frame.setFixedHeight(681)
-
-                # Disabling the toggler btn
-                self.DisplayReschedule_Btn.setEnabled(False)
-                self.DisplayReschedule_Btn.setStyleSheet("QPushButton {\n"
-"    border-image: none;\n"
-"    background-color: rgba(110, 204, 175, .2);\n"
-"    color: rgba(0, 0, 0, .2);\n"
-"    border: 2px solid rgba(0, 0, 0, .2);\n"
-"}\n"
-"\n"
-"QPushButton::hover {\n"
-"    background-color: rgb(139, 231, 100);\n"
-"    color: white;\n"
-"}")
-
-
+        
+        self.load_nav()
 
 
         #define widgets
-        self.logoutPushButton = self.findChild(QPushButton, "Nav_LogoutBtn")
-        self.appointmentsPushButton = self.findChild(QPushButton, "Nav_Appointments")
-        self.checkinPushButton = self.findChild(QPushButton, "Nav_CheckinBtn")
-        self.checkoutPushButton = self.findChild(QPushButton, "Nav_CheckoutBtn")
-        self.makeReferralPushButton = self.findChild(QPushButton, "Nav_MakeReferralBtn")
-        self.labOrdersPushButton = self.findChild(QPushButton, "Nav_LabOrdersBtn")
-        self.approveAppointmentsPushButton = self.findChild(QPushButton, "Nav_ApproveAppointmentsBtn")
 
+        # Appiontment Nav 
         self.newPatientPushButton = self.findChild(QPushButton, "NewPatient_Btn")
         self.reschedulingPushButton = self.findChild(QPushButton, "DisplayReschedule_Btn")
         self.makeSchedulePushButton = self.findChild(QPushButton, "DisplaySchedule_Btn")
@@ -185,29 +69,13 @@ class UI(QMainWindow):
         self.CA_SearchForAppointmentsPushButton = self.findChild(QPushButton, "SearchForAppointments_Btn_CA")
         self.CA_CancelAppointmentPushButton = self.findChild(QPushButton, "CancelAppointment_Btn_CA")
         self.CA_ClearInputsPushButton = self.findChild(QPushButton, "CA_ClearInputsBtn")
-        self.currentUserLabel = self.findChild(QLabel, "currentUserLabel")
-
-        # Displaying current user in label and title
-        self.currentUserLabel.setText("")
-        self.currentUserLabel.setText("Current User: " + currentUsername[0])
-        self.setWindowTitle("")
-        self.setWindowTitle("Forsyth Family Practice Center - Scheduling Appointments -|- User: " + currentUsername[0])
-
-        #Do something
-        self.logoutPushButton.mousePressEvent = lambda event: logoutUser(self)
-
-        self.checkinPushButton.mousePressEvent = lambda event: enterCheckInWindow(self)
-        self.checkoutPushButton.mousePressEvent = lambda event: enterCheckOutWindow(self)
-        self.makeReferralPushButton.mousePressEvent = lambda event: enterMakeReferralWindow(self)
-        self.labOrdersPushButton.mousePressEvent = lambda event: enterLabOrdersWindow(self)
-        self.approveAppointmentsPushButton.mousePressEvent = lambda event: enterAppointmentApproveViaPortalWindow(self)
 
 
-
+        # Events for buttons
         self.newPatientPushButton.mousePressEvent = lambda event: enterNewPatientWindow(self)
-        self.reschedulingPushButton.clicked.connect(displayRescheduleAppointmentFrame)
-        self.makeSchedulePushButton.clicked.connect(displayInputsFrame)
-        self.cancelPushButton.clicked.connect(displayCancelAppointmentFrame)
+        self.reschedulingPushButton.clicked.connect(self.displayRescheduleAppointmentFrame)
+        self.makeSchedulePushButton.clicked.connect(self.displayInputsFrame)
+        self.cancelPushButton.clicked.connect(self.displayCancelAppointmentFrame)
 
         self.SA_PatientDOBDateEdit.setDate(QDate.currentDate())
         self.SA_AppointmentDateDateEdit.setDate(QDate.currentDate())
@@ -234,31 +102,10 @@ class UI(QMainWindow):
         self.CA_CancelAppointmentPushButton.mousePressEvent = lambda event: cancelAppointment(self)
 
 
-        greyOutReferralsAndLabOrdersForPhysicians(self)
-        #Hide the app
-        # self.hide()
-
-
-
-
-    # This will make it so when the user clicks the red x, it closes the app
-    def closeEvent(self, event):
-        sys.exit()
-
-    def moveEvent(self, event):
-        prevWindowCoords.clear()
-
-        coords = self.pos()
-
-        prevWindowCoords.append(coords.x())
-        prevWindowCoords.append(coords.y())
-
-        print(prevWindowCoords)
-
 
 #initializing app
 app = QApplication(sys.argv)
 UIWindow = UI()
-
-
-# app.exit()
+if __name__ == "__main__":
+    UIWindow.show()
+    app.exec_()
